Fix Move File To Trash: Use build it solution
