+++ src/Native/Windows.cs
@@ -33,6 +33,19 @@ namespace SourceGit.Native
             public int cyBottomHeight;
         }
 
+        [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Auto)]
+        internal struct SHFILEOPSTRUCT
+        {
+            public IntPtr hwnd;
+            public uint wFunc;
+            public string pFrom;
+            public string pTo;
+            public ushort fFlags;
+            public bool fAnyOperationsAborted;
+            public IntPtr hNameMappings;
+            public string lpszProgressTitle;
+        }
+
         [DllImport("dwmapi.dll")]
         private static extern int DwmExtendFrameIntoClientArea(IntPtr hwnd, ref MARGINS margins);
 
@@ -47,6 +60,9 @@ namespace SourceGit.Native
 
         [DllImport("shell32.dll", CharSet = CharSet.Unicode, SetLastError = false)]
         private static extern int SHOpenFolderAndSelectItems(IntPtr pidlFolder, int cild, IntPtr apidl, int dwFlags);
+
+        [DllImport("shell32.dll", CharSet = CharSet.Auto)]
+        private static extern int SHFileOperation(ref SHFILEOPSTRUCT FileOp);
 
         [DllImport("user32.dll")]
         private static extern bool GetWindowRect(IntPtr hwnd, out RECT lpRect);
@@ -253,6 +269,28 @@ namespace SourceGit.Native
             Process.Start(start);
         }
 
+        public bool MoveFileToTrash(string file)
+        {
+            SHFILEOPSTRUCT fileOp = new SHFILEOPSTRUCT
+            {
+                wFunc = FO_DELETE,
+                pFrom = file + '\0',
+                pTo = null,
+                fFlags = (ushort)(FOF_ALLOWUNDO | FOF_NOCONFIRMATION),
+                fAnyOperationsAborted = false,
+                hNameMappings = IntPtr.Zero,
+                lpszProgressTitle = null
+            };
+
+            // Move the file in the Trash
+            bool result = SHFileOperation(ref fileOp) == 0;
+
+            // If move ok delete the temp file
+            if (result) File.Delete(file);
+
+            return result;
+        }
+
         private void FixWindowFrameOnWin10(Window w)
         {
             // Schedule the DWM frame extension to run in the next render frame
@@ -431,5 +469,9 @@ namespace SourceGit.Native
 
             return null;
         }
+
+        private const uint FO_DELETE = 0x0003;    // Delete operation
+        private const uint FOF_ALLOWUNDO = 0x0040; // Put in Trash
+        private const uint FOF_NOCONFIRMATION = 0x0010; // No confirmation
     }
 }
